Respect updated DNS SD broadcast names
When a device is advertising over dns-sd and the TX bits change (device
name) a message is sent out to indicate that the old adverisment should
be expired and the new one should be used.

In the current implementation of the casting app, we stop listening to
updates after the `discoveryDurationSeconds` timeout. This causes us to
miss messages and Android seems to fall back to the cached value of the
DNS advertisment.

We need to change the usage pattern here so that clients stop the
discovery manually after it is started and keep it active for the
lifetime they need it for.

I also updated the test apps to behave a bit nicer with discovered
commissioners.