Default to use an IPV4 address (#114)
* Default to use an IPV4 address

On some devices it appears that the commissioner may not have an
appropriate IPV6 address and attempts to send a user directed
commissioning request to a device using its IPV6 address.

Its a bit unclear how this situation happens but in theory it could
occur. In most cases this should be safe to default to an ipv4 address
if it is available.

The implementation of the user directed commissioning server does
something similar.