Take connection operations off the UI thread (#122)
Clicking connect in the tv-casting-app freezes the UI for a few seconds
because it tries to attempt the connection on the UI thread. This puts
it in an async thread so that the UI is not blocked while its trying to
establish that connection.

Tested by attempting the connection and verifying it wasn't paused and
still showed the right messages and connected successfully.