Use TargetServiceInfo to resolve the ServiceType
On my network, the service type resolved with an extra `.` at the front
(`._matterc._udp`) even though the requested service type was
`_matterc._udp`. This is impactful because internally we do a check for
the length of the service type and it has to match the max length (13).
In this case because of the extra "." the check fails.

I've changed this to just use the ServiceInfo we used to start the
resolution with so that there is no contention here.