Enable TV Casting App Example to using Hardware-backed key generation for DAC
It is best practice for private keys to not be available to the application layer in plain form. In
fact you can generate keys within the secure element on device, or TrustZone, that is not extractable
in plain form. In order to support such best practices, the change addresses this by modifying the DAC Provider's
SignWithDeviceAttestationKey function that is provided by the example tv castin app to:

- For Android, ask the Java application layer to do the signing via the Android Keystore.
- For iOS, request the SecKeyRef for the private key as part of setting the DAC provider and hence allowing the SDK
 to just sign using the Security APIs.