Update JNI Interface in tv-app to allow app to sign messages for DAC stored in Android KeyStore
In an effort to leverage hardware keys for DAC, the JNI layer was updated to ask the JVM layer to
sign message by leveraging the Android KeyStore instead of requesting to the raw private key bytes
only to sign it within the c++ layer. This adds a level of security to the device attestation process.